requirements.txt
Django>=5.0.0
psycopg2-binary>=2.9.9
celery>=5.3.6
redis>=5.0.1
openai>=1.12.0
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
djangorestframework>=3.15.1 